+++ packages/dbx-firebase/src/lib/model/system/systemstate.document.store.directive.ts
@@ -1,0 +1,14 @@
+import { Directive } from '@angular/core';
+import { SystemState, SystemStateDocument, SystemStateStoredData } from '@dereekb/firebase';
+import { DbxFirebaseDocumentStoreDirective, provideDbxFirebaseDocumentStoreDirective } from '../store/store.document.directive';
+import { SystemStateDocumentStore } from './systemstate.document.store';
+
+@Directive({
+  selector: '[dbxFirebaseSystemStateDocument]',
+  providers: provideDbxFirebaseDocumentStoreDirective(DbxFirebaseSystemStateDocumentStoreDirective, SystemStateDocumentStore)
+})
+export class DbxFirebaseSystemStateDocumentStoreDirective<T extends SystemStateStoredData = SystemStateStoredData> extends DbxFirebaseDocumentStoreDirective<SystemState<T>, SystemStateDocument<T>, SystemStateDocumentStore<T>> {
+  constructor(store: SystemStateDocumentStore<T>) {
+    super(store);
+  }
+}